Fixing up schemas and resources
